refactor(Name.java) : VO 패키지 생성
