+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.common.Name;
+import kitchenpos.common.vo.Name;
 
 import javax.persistence.*;
 import java.util.Objects;
