+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.common.Quantity;
+import kitchenpos.common.vo.Quantity;
 
 import javax.persistence.*;
 import java.util.Objects;
