+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.domain;
 
-import kitchenpos.common.Name;
-import kitchenpos.common.Price;
+import kitchenpos.common.vo.Name;
+import kitchenpos.common.vo.Price;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
