+++ src/test/java/kitchenpos/common/QuantityTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.common;
 
+import kitchenpos.common.vo.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
