+++ src/test/java/kitchenpos/common/fixture/PriceFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.common.fixture;
 
-import kitchenpos.common.Price;
+import kitchenpos.common.vo.Price;
 
 import java.math.BigDecimal;
