+++ src/test/java/kitchenpos/order/domain/fixture/OrderLineItemFixture.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.domain.fixture;
 
-import kitchenpos.common.Name;
-import kitchenpos.common.Price;
-import kitchenpos.common.Quantity;
+import kitchenpos.common.vo.Name;
+import kitchenpos.common.vo.Price;
+import kitchenpos.common.vo.Quantity;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
